Patching GPP 'configure' to trust uname -p
Relates to #2

The GPP's provided node generation script ignores the value of
uname -p if it doesn't match a very narrow set of dictionary KV
pairs and then slams the result to x86 like that's a good idea.
This set of patches amends this behavior to instead trust what
the processor is telling it.