Shifted inclusion of .so into main packages
REDHAWK's use of .so files goes against conventions per Debian, etc.
in that ComponentHost, for example, directly looks for the .so
(link or file) or else will fail.  Similarly entrypoints for projects
defined as Shared Libraries, the entrypoint is an .so, etc.  So it's
simply easier to suppress (dev-so) and include rather than induce
significant breakage by re-writing or patching this out.

Signed-off-by: Thomas Goodwin <btgoodwin@geontech.com>