Fixing redhawk-waveform.bbclass to parse and work correctly.
This addresses #48 by both finding the waveform SAD and then
installing it without being executable.

This addresses #45 by removing the EXPORT_FUNCTIONS use, which
fails to parse since the class name has hyphens in it.  Rather
than potentially break a lot of recipes by renaming all classes
to match the required change (redhawk_waveform.bbclass), we're
opting to not export the install task since as long as it works,
there really only is one way to install a waveform.

Signed-off-by: Thomas Goodwin <btgoodwin@geontech.com>