Extended GPP PRF patching to include os_name
When launching components, the GPP PRF is parsed and the values
are read from it rather than the actual properties.  Unlike
processor_name, the os_name is actually required to be set
because every component has that property maching set in its
implementation.  Therefore we set the value to the default,
Linux.

Signed-off-by: Thomas Goodwin <btgoodwin@geontech.com>