Zeus support
layer.conf updated for zeus support along with giflib and python-numpy
(the latter of which isn't in meta-python2, for some reason, but is
required).

burstiointerfaces and gpp had fuzz warnings for patches that were
addressed

rh-fftlib required a patch for its vector allocator.

redhawk-python required an explicit RDEPENDS against python now
rather than an inferred one by all the python modules on which
it relies.

Signed-off-by: B. Thomas Goodwin <btgoodwin@geontech.com>