layer dependency order is incorrect
This won't actually break anything if using the sample conf
directly, but if using 'bitbake-layers add-layer' and using
this sample as a guide, it will fail because networking
depends on a layer after it in the list, which will cause
a parsing error once networking is added.  You can get around
it by using -f with that command, but just for the sake of
clarity in this being a form of documentation: here's the right
order.

Signed-off-by: B. Thomas Goodwin <btgoodwin@geontech.com>