redhawk sandbox: reverse dependency load order
It's unclear why this is necessary with Yocto builds, but the gist
is that if Component A depends on softpkg B, which depends on
softpkg C, in the sandbox, you must load C before B in order for
A to load even though the dlopen() calls are set to WEAK and work
fine otherwise.  The only core asset example case is rh.DataConverter.
All other components list the dsp softpkg first, then the fftlib
softpkg (which then points to dsp).  One could make this change to
the rh.DataConverter.spd.xml as well, but loading dependencies from
the tail first seems to be the working, general solution.

Signed-off-by: B. Thomas Goodwin <btgoodwin@geontech.com>