oeqa: removed domain-init requirement
Some tests were overloading the gpp/domain that were running
on init, and then causing other tests to fail.  This is the
start of some re-work to make this test follow a more
conventional setup/tear-down pattern each time it's run by
providing its own unique domain and GPP instances (assuming
a node bearing a GPP is installed, more on that in the next
commits).

Signed-off-by: B. Thomas Goodwin <btgoodwin@geontech.com>